\name{surrosurv-package}
\alias{surrosurv-package}
\docType{package}
\title{
\packageTitle{surrosurv}
}
\description{
\packageDescription{surrosurv}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{surrosurv}
\packageIndices{surrosurv}
}
\author{
\packageAuthor{surrosurv}

Maintainer: \packageMaintainer{surrosurv}
}
\references{
Rotolo F, Paoletti X, Burzykowski T, Buyse M, Michiels S.
Evaluation of failure time surrogate endpoints in individual patient data meta-analyses of randomized clinical trials. A Poisson approach.
\emph{Working paper}.

Burzykowski T, Molenberghs G, Buyse M et al.
Validation of surrogate end points in multiple randomized clinical trials with failure time end points.
\emph{Journal of the Royal Statistical Society C} 2001;
\bold{50}:405--422.
\doi{10.1111/1467-9876.00244}.

Gasparrini A, Armstrong B, Kenward MG.
Multivariate meta-analysis for non-linear and other multi-parameter associations.
\emph{Statistics in Medicine} 2012; \bold{31}:3821--39.
\doi{10.1002/sim.5471}

Burzykowski T, Molenberghs G, Buyse M (2005). 
emph{The Evaluation of Surrogate Endpoints}. Springer, New York.
}
\keyword{ surrogate endpoint }
\keyword{ survival }
\keyword{ meta-analysis }
\keyword{ randomized controlled trial }
\keyword{ copula }
\keyword{ Clayton }
\keyword{ Plackett }
\keyword{ Hougaard }
\keyword{ proportional hazard model }
\keyword{ generalized linear mixed model }
\keyword{ Poisson }

\seealso{
\CRANpkg{Surrogate}, \CRANpkg{mvmeta}
}